chore: :recycle: Move AccountFormatters methods
