+++ lib/ui/util/transfer_recipient_formatters.dart
@@ -1,6 +1,6 @@
-import 'package:aewallet/ui/util/account_formatters.dart';
 import 'package:aewallet/ui/util/address_formatters.dart';
 import 'package:aewallet/ui/views/transfer/bloc/state.dart';
+import 'package:aewallet/util/account_formatters.dart';
 import 'package:flutter_gen/gen_l10n/localizations.dart';
 
 extension TransferRecipientFormatters on TransferRecipient {
