+++ lib/ui/views/add_account/bloc/provider.dart
@@ -1,5 +1,4 @@
 /// SPDX-License-Identifier: AGPL-3.0-or-later
-import 'package:aewallet/application/account/account_notifier.dart';
 import 'package:aewallet/application/account/accounts_notifier.dart';
 import 'package:aewallet/application/session/session.dart';
 import 'package:aewallet/application/transaction_repository.dart';
@@ -8,6 +7,7 @@ import 'package:aewallet/domain/models/transaction.dart';
 import 'package:aewallet/model/data/account.dart';
 import 'package:aewallet/ui/util/transaction_send_event_error_localization.dart';
 import 'package:aewallet/ui/views/add_account/bloc/state.dart';
+import 'package:aewallet/util/account_formatters.dart';
 import 'package:archethic_lib_dart/archethic_lib_dart.dart' as archethic;
 import 'package:event_taxi/event_taxi.dart';
 import 'package:flutter/material.dart';
