+++ lib/ui/views/rpc_command_receiver/get_accounts/command_handler.dart
@@ -1,8 +1,8 @@
-import 'package:aewallet/application/account/account_notifier.dart';
 import 'package:aewallet/domain/models/core/result.dart';
 import 'package:aewallet/domain/rpc/command_dispatcher.dart';
 import 'package:aewallet/domain/rpc/commands/command.dart';
 import 'package:aewallet/infrastructure/datasources/account.hive.dart';
+import 'package:aewallet/util/account_formatters.dart';
 import 'package:archethic_wallet_client/archethic_wallet_client.dart' as awc;
 
 class GetAccountsCommandHandler extends CommandHandler {
