+++ lib/ui/views/transfer/layouts/transfer_sheet.dart
@@ -8,7 +8,6 @@ import 'package:aewallet/model/primary_currency.dart';
 import 'package:aewallet/modules/aeswap/domain/models/dex_token.dart';
 import 'package:aewallet/ui/themes/archethic_theme.dart';
 import 'package:aewallet/ui/themes/styles.dart';
-import 'package:aewallet/ui/util/account_formatters.dart';
 import 'package:aewallet/ui/util/amount_formatters.dart';
 import 'package:aewallet/ui/util/formatters.dart';
 import 'package:aewallet/ui/util/ui_util.dart';
@@ -21,6 +20,7 @@ import 'package:aewallet/ui/widgets/components/app_text_field.dart';
 import 'package:aewallet/ui/widgets/components/paste_icon.dart';
 import 'package:aewallet/ui/widgets/dialogs/accounts_dialog.dart';
 import 'package:aewallet/ui/widgets/tokens/verified_token_icon.dart';
+import 'package:aewallet/util/account_formatters.dart';
 import 'package:aewallet/util/currency_util.dart';
 import 'package:aewallet/util/user_data_util.dart';
 import 'package:archethic_dapp_framework_flutter/archethic_dapp_framework_flutter.dart'
